+++ es-design-system/pages/atoms/icons.vue
@@ -337,6 +337,10 @@
                 <code>IconTree</code>
             </li>
             <li>
+                <icon-upload />
+                <code>IconUpload</code>
+            </li>
+            <li>
                 <icon-verified />
                 <code>IconVerified</code>
             </li>
@@ -395,12 +399,24 @@
                 <code>IconCommunitySolar</code>
             </li>
             <li>
+                <icon-contact-us />
+                <code>IconContactUs</code>
+            </li>
+            <li>
                 <icon-electric-vehicle />
                 <code>IconElectricVehicle</code>
             </li>
             <li>
                 <icon-heat-pump />
                 <code>IconHeatPump</code>
+            </li>
+            <li>
+                <icon-info-blog-post />
+                <code>IconInfoBlogPost</code>
+            </li>
+            <li>
+                <icon-marketplace />
+                <code>IconMarketplace</code>
             </li>
             <li>
                 <icon-rooftop-solar />
@@ -436,6 +452,29 @@
             <li>
                 <icon-youtube />
                 <code>IconYoutube</code>
+            </li>
+        </ul>
+
+        <h2>
+            File Icons
+        </h2>
+        <p>
+            Default size for file icons is 56px by 73px.
+        </p>
+        <ul
+            class="ds-icon-list m-0 mb-300 p-0"
+            :class="{ [textColorClass]: true }">
+            <li>
+                <icon-file-doc />
+                <code>IconFileDoc</code>
+            </li>
+            <li>
+                <icon-file-docx />
+                <code>IconFileDocx</code>
+            </li>
+            <li>
+                <icon-file-pdf />
+                <code>IconFilePdf</code>
             </li>
         </ul>