+++ es-vue-base/src/lib-components/EsNavBarTopLevelMenu.vue
@@ -125,7 +125,8 @@
                             :link="featuredArticle.linkUrl"
                             :new-tab="featuredArticle.newTab"
                             :name="featuredArticle.lede"
-                            :image500="featuredArticle.image500" />
+                            :image500="featuredArticle.image500"
+                            :image-alt="featuredArticle.imageAlt" />
                     </b-row>
                 </b-container>
             </div>